* Compiles locally again using source versions of Elypiai, Commandler, and ComCord * Updated README for more organization. * Replaced name from Syed Shah to Seth Falco * Adding all code back into Alexis (WIP) * Cleaned up and renamed all configuration. * Updated for new Commandler/Elypiai versions, including changing method calls. * Migrated most `Controllers` over, 10 still remain. * Migrated all listeners over and split them up into more smaller listeners. * Further implementing Messengers for various types of objects to ensure there is a clean non-embed fall back. * Renamed Module to Controller * Created DatabaseUtils (may rename to HibernateUtils) to store generic utilities for database. * Cleaned up Main file significantly, now initialized and loads succesfully with controllers and listeners. * Moved from JPA configuration, to hibernate configuration (for easier programmatic setting) * Replaced `...` with actual elipsis character. * Replaced all annoations with XML file. [TEMP] * Added dependency on twitch4j (Twitch Library for Java) * Added test cases to verify YouTubeUtils functions work correctly.

+++ src/main/java/org/elypia/alexis/entities/GuildFeature.java
@@ -21,6 +21,9 @@ package org.elypia.alexis.entities;
 import javax.persistence.*;
 import java.util.Date;
 
+/**
+ * @author seth@elypia.org (Seth Falco)
+ */
 @Entity(name = "guild_feature")
 @Table
 public class GuildFeature {
@@ -31,7 +34,7 @@ public class GuildFeature {
     private int id;
 
     @Column(name = "guild_id")
-    private int guildId;
+    private Long guildId;
 
     @Column(name = "feature_name")
     private String name;
@@ -49,11 +52,11 @@ public class GuildFeature {
         return id;
     }
 
-    public int getGuildId() {
+    public Long getGuildId() {
         return guildId;
     }
 
-    public void setGuildId(int guildId) {
+    public void setGuildId(Long guildId) {
         this.guildId = guildId;
     }
